Task records. Plain text database.
Instead of using Fossil and all the learning and fighting I'd have to
do with that, all I really wanted was a task tracker. I don't want to
lean on Github for all of this stuff, so I did it myself with
`recutils` and Emacs.

FossilOrigin-Name: b4ee02ae029ef21e456ddc25223051946c67b13d1f5a480aeec13601797bd44c